Fixed typing in variwide series.
