+++ js/Series/Variwide/VariwideComposition.js
@@ -14,6 +14,11 @@ import Tick from '../../Core/Axis/Tick.js';
 import Axis from '../../Core/Axis/Axis.js';
 import U from '../../Core/Utilities.js';
 var addEvent = U.addEvent, wrap = U.wrap;
+/* *
+ *
+ * Composition
+ *
+ * */
 Tick.prototype.postTranslate = function (xy, xOrY, index) {
     var axis = this.axis, pos = xy[xOrY] - axis.pos;
     if (!axis.horiz) {
