+++ js/Series/Variwide/VariwidePoint.js
@@ -22,17 +22,13 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import ColumnSeries from '../Column/ColumnSeries.js';
+import BaseSeries from '../../Core/Series/Series.js';
+var ColumnSeries = BaseSeries.seriesTypes.column;
 import U from '../../Core/Utilities.js';
 var isNumber = U.isNumber;
 /* *
  *
  * Class
- *
- * */
-/* *
- *
- * VariwidePoint class
  *
  * */
 var VariwidePoint = /** @class */ (function (_super) {
@@ -44,6 +40,7 @@ var VariwidePoint = /** @class */ (function (_super) {
          * Properites
          *
          * */
+        _this.crosshairWidth = void 0;
         _this.options = void 0;
         _this.series = void 0;
         return _this;
